refactor(agents): extract bootstrap warning contracts
